Avoid unnecessary cast
git-svn-id: file:///home/roberto/Git/svn@6242 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: beb1443684ba55c770b3c98df492c7fde0e40b4b

+++ trunk/xmipp/libraries/data/multidim_array.h
@@ -3085,11 +3085,11 @@ public:
         computeStats(avg0, stddev0, minval, maxval);
 
         if (stddev0 != 0)
-            a = static_cast< double >(stddevF) / static_cast< double >(stddev0);
+            a = stddevF / stddev0;
         else
             a = 0;
 
-        b = static_cast< double >(avgF) - a * static_cast< double >(avg0);
+        b = avgF - a * avg0;
 
         T* ptr=NULL;
         unsigned long int n;
